% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sicomore-package.R, R/sicomore.R
\docType{package}
\name{sicomore}
\alias{sicomore}
\title{sicomore}
\usage{
sicomore(
  y,
  X.list,
  compressions = rep("mean", length(X.list)),
  selection = c("rho-sicomore", "sicomore", "mlgl"),
  cuts = lapply(X.list, function(X) unique(round(10^seq(log10(ncol(X)), log10(2), len =
    100)))),
  choice = c("lambda.min", "lambda.min"),
  method.clus = c("ward.D2", "ward.D2"),
  depth.cut = c(3, 3),
  mc.cores = 1,
  taxonomy = NULL,
  verbose = TRUE,
  stab = TRUE,
  stab.param = list(B = c(100, 100), cutoff = c(0.75, 0.75), PFER = c(1, 1))
)
}
\arguments{
\item{y}{response variable (phenotype)}

\item{X.list}{a list of input matrices. Must have the same number of rows (number of observations). May have different number of columns (predictors)}

\item{compressions}{a vector of string for the compression methods for each data sets. Default used the mean to compressed predictors in the same group.}

\item{selection}{method used to perform variable selection. Either 'sicomore', 'mlgl' or 'rho-sicomore' (see details). Default is 'sicomore'.}

\item{cuts}{a list of numeric vector defining the cut levels to be considered for each hierarchy. By default a sequence of 100 levels is used.}

\item{choice}{a string (either "lambda.min" or "lambda.1se"). Indicates how the tuning parameter is chosen in the penalized regression approach.}

\item{method.clus}{a vector of string specifying the method used for the hierarchical clustering, one for each input matrix in X.list.
By default, the hierarchy is obtain by a WARD clustering on the scaled input matrix.
To use an SNP-specific spatially contrained hierarchical clustering \insertCite{dehman}{sicomore} from package adjclust, specify "snpClust".
It is also possible to specify no hierarchy with "noclust".}

\item{depth.cut}{a vector of integers specifying the depth of the search space for the variable selection part of the algorithm.
This argument allows to increase the speed of the algorithm by restraining the search space without affecting too much the performance.
A value between 3 and 6 is recommended, the smaller the faster.}

\item{mc.cores}{an integer for the number of cores to use in the parallelization of the cross-validation and some other functions. Default is 1.}

\item{taxonomy}{a hierarchical tree object constructed using taxonomical unit.
This argument could be useful if one want to provid a specific hierarchical tree based on taxonomy rather than euclidean distance.
Recommended for those who want to detect genomic-metagenomic interactions.}

\item{verbose}{not yet documented}

\item{stab}{not yet documented}

\item{stab.param}{not yet documented}
}
\value{
an RC object with class \code{sicomore} with methods \code{plot()}, \code{getSignificance()} and the following fields
\itemize{
 \item{pval:}{A matrix of p-values for each interactions effects between the compressed variables originating from 2 input matrices.}
 \item{pval.beta1:}{A matrix of p-values for the corresponding main effects of the compressed variables originating from the first input matrix}
 \item{pval.beta2:}{A matrix of p-values for the corresponding main effects of the compressed variables originating from thesecond input matrix}
 \item{models}{A class'sicomore-model' RC object obtain from \code{getHierLevel} function and with methods \code{nGrp()}, \code{nVar()}, \code{getGrp()}, \code{getVar()},
 \code{getCV()}, \code{getX.comp()}, \code{getCoef()}}
 \item{tuplets:}{A list of integer vector specifying the indexes of the compressed variables which are fitted together in a linear model with interaction.}
 \item{dim:} A array of integers specifying the dimension of the compressed matrices.
}
}
\description{
Selection of Interaction effects in COmpressed  Multiple Omics REpresentation

Selection of Interaction effects in COmpressed  Multiple Omics REpresentation
}
\details{
From a set of input matrices and phenotype related to the same set of individual, sicomore is a two-step method which
1. find and select groups of correlated variables in each input matrix which are good predictors for the common phenotype
2. find the most predictive interaction effects between the set of data by testing for interaction between the selected groups of each input matrix

The methods for variable selection are variants of Lasso or group-Lasso designed to perform selection of interaction between multiple hierarchies:
'sicomore' and 'rho-sicomore' (see \insertCite{sicomore;textual}{sicomore}) use a LASSO penalty on compressed groups of variables along the hierarchies to select interactions.
rho-sicomore is a variant where a more sound weighting scheme is used dependending on the level of the hierarchy considered. The method 'mlgl' of \insertCite{grimonprez_PhD;textual}{sicomore}
uses a group-Lasso penalty which does not require compression but requires heavier computational resources.
}
\references{
\insertAllCited{}
}
\author{
Julien Chiquet \email{julien.chiquet@inrae.fr}

Florent Guinot \email{fly.guinot@gmail.com}

Marie Szafranski  \email{marie.szafranski@math.cnrs.fr}

Christophe Ambroise \email{christophe.ambroise@genopole.cnrs.fr}
}
